vtep-emulator: Support L3 Logical Router
This patch is to support L3 DLR E-W traffic functionality in ovs-vtep. So it is
focus on supporting switch-binding field in Logical_Router Table schema.

Like what we do for logical switch, there are 2 parts of work:

1> physical handling
    In this part, create binding between logical router and logical switch,
    create peer ports between them, and create related openflow tables,
    similiar with logical switch
2> run function
    In this part, update lif mac-ip bindings if any change, create ARP table to
    respond arp request for gateway,  then update routing table and forwarding
    table to create openflow entry to update src and dst macs in the packet.

Another part of work is to enhance vtep-ctl to let it recognize Logical Router
Table in VTEP DB, additionally, support add-lr command, but it could be trival.

There are some limiations and assumptions here
1>  single-tier logical router only
2>  don't support dynamically arping destination VM's mac when packet arrive at
    gateway, so ipaddr field in local/remote mac table need to be filled by advance.
3>  E-W traffic only

Signed-off-by: Kyle Mestery <mestery@mestery.com>