While commit ab39371d68842b7e4000cc5d8718e6fc04e92795 (ovn-controller: Handle physical changes correctly) addressed unit test failures, it did so at the cost of performance: [1] notes that ovn-controller cpu usage is now pegged at 100%.
Root cause of this is that while the storage for tunnels is
persisted, their creation is not (which the above changed
incorrectly assumed was the case).  This patch persists
tunneled data across invocations of physical_run.  A side
effect is that renaming of localfvif_map_changed variable
to physical_map_changed and extending its scope to include
tunnel changes.

[1] http://openvswitch.org/pipermail/dev/2016-July/076058.html

Signed-off-by: Ryan Moats <rmoats@us.ibm.com>
Signed-off-by: Kyle Mestery <mestery@mestery.com>