This patchset mimics the changes introduced in
f199df26 (ovsdb-idl: Add partial map updates functionality.)
010fe7ae (ovsdb-idlc.in: Autogenerate partial map updates functions.)
7251075c (tests: Add test for partial map updates.)

but for columns that store sets of values rather than key-value
pairs.  These columns will now be able to use the OVSDB mutate
operation to transmit deltas on the wire rather than use
verify/update and transmit wait/update operations on the wire.

This patch is marked as an RFC as it is still necessary to
update the python bindings to use this and partial map.
Once those are done, this will be reissued as part one of a
two part patch set.

Signed-off-by: Ryan Moats <rmoats@us.ibm.com>
Signed-off-by: Kyle Mestery <mestery@mestery.com>